# packages/contracts/openapi/control-api.yaml
openapi: 3.0.3
info:
  title: KofClub Control API
  version: 0.1.0
  description: |
    Control-plane HTTP API (accounts, clubs, roles, table configuration,
    virtual-chip ledger administration, hand history).

    Conventions (see docs/adr/ADR-009-wire-conventions.md):
    - All endpoints are versioned under `/v1` (health probes are unversioned).
    - JSON bodies use camelCase; enums are UPPERCASE strings.
    - Identifiers are UUID strings (UUIDv7 preferred).
    - Timestamps are UTC RFC 3339 strings.
    - Chip amounts are integers (int64). Floating point is never used.
    - Every response carries an `X-Request-Id` header; clients may send one.
    - State-changing requests that may be retried accept an `Idempotency-Key` header.
    - Errors always use the `ErrorEnvelope` shape with a machine-readable `code`.

    Virtual chips have no monetary value. This API intentionally contains no
    deposit, withdrawal, cash-out or payment functionality (ADR-006).
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /v1/clubs:
    post:
      tags: [clubs]
      operationId: createClub
      summary: Create a club (caller becomes OWNER)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClubRequest'
      responses:
        '201':
          description: Club created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          $ref: '#/components/responses/Error'
    get:
      tags: [clubs]
      operationId: listMyClubs
      summary: Clubs I am an active member of
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Clubs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubList'
  /v1/clubs/join:
    post:
      tags: [clubs]
      operationId: joinClubByCode
      summary: Join a club with a club join code or an invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinClubRequest'
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [clubs]
      operationId: getClub
      summary: Club details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [clubs]
      operationId: updateClub
      summary: Rename the club or edit its description (OWNER)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClubRequest'
      responses:
        '200':
          description: Updated club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/transfer-ownership:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [clubs]
      operationId: transferClubOwnership
      summary: Hand the club to another active member (OWNER); the previous owner becomes ADMIN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferOwnershipRequest'
      responses:
        '200':
          description: Club after the transfer (viewed by the previous owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/join:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [clubs]
      operationId: joinClub
      summary: Join this club by join code or invite code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinClubRequest'
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/leave:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [clubs]
      operationId: leaveClub
      summary: Leave the club (owners cannot leave)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Left
        '403':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/join-code/rotate:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [clubs]
      operationId: rotateClubJoinCode
      summary: Replace the club join code (ADMIN+)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Club with the new join code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '403':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/members:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [clubs]
      operationId: listClubMembers
      summary: Members (keyset-paginated)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          required: false
          description: Filter by status (staff only; members always see ACTIVE).
          schema:
            $ref: '#/components/schemas/MemberStatus'
      responses:
        '200':
          description: Page of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '403':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [clubs]
      operationId: updateClubMember
      summary: Change a member's role or status (ban/unban)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: Updated member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/invites:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    post:
      tags: [clubs]
      operationId: createClubInvite
      summary: Create an invite code (AGENT+; AGENT-role invites need ADMIN+)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created; the code is only returned once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteCreated'
        '403':
          $ref: '#/components/responses/Error'
    get:
      tags: [clubs]
      operationId: listClubInvites
      summary: List invites (codes are never returned)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteList'
        '403':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/invites/{inviteId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
      - name: inviteId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    delete:
      tags: [clubs]
      operationId: revokeClubInvite
      summary: Revoke an invite
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Revoked
        '404':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/audit-log:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [clubs]
      operationId: listClubAuditLog
      summary: Privileged-action audit log (ADMIN+, platform admins)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of audit records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/me/hands:
    get:
      tags: [history]
      operationId: listMyHands
      summary: Hands I was dealt into (finished hands, newest first)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of hand summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandSummaryPage'
        '401':
          $ref: '#/components/responses/Error'
  /v1/hands/{handId}:
    parameters:
      - $ref: '#/components/parameters/HandId'
    get:
      tags: [history]
      operationId: getHand
      summary: Visible hand record (ADR-008 visibility policy)
      description: |
        Participants see the public action log, the board, cards shown at
        showdown and their own hole cards. Club staff (ADMIN+) and platform
        admins see the same public record without anyone's unrevealed cards.
        Anyone else receives HAND_NOT_FOUND.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Hand
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandDetail'
        '404':
          $ref: '#/components/responses/Error'
  /v1/clubs/{clubId}/hands:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [history]
      operationId: listClubHands
      summary: Finished hands in a club (ADMIN+ audit view, newest first)
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Uuid'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of hand summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandSummaryPage'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/admin/clubs/{clubId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    patch:
      tags: [admin]
      operationId: adminSetClubStatus
      summary: Suspend (view-only) or reinstate a club (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminClub'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    HandId:
      name: handId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    ClubId:
      name: clubId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous page's `nextCursor`.
      schema:
        type: string
        maxLength: 512
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key making a retried state-changing request a no-op.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    Error:
      description: Machine-readable error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp
    ChipAmount:
      type: integer
      format: int64
      description: Integer amount of virtual chips (no monetary value).
...
    ClubRole:
      type: string
      enum: [OWNER, ADMIN, AGENT, MEMBER]
    ClubStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, CLOSED]
    MemberStatus:
      type: string
      enum: [ACTIVE, BANNED, LEFT]
    Club:
      type: object
      additionalProperties: false
      required: [id, name, description, status, ownerUserId, createdAt, myRole]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ClubStatus'
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        myRole:
          type: string
          enum: [OWNER, ADMIN, AGENT, MEMBER, null]
          nullable: true
          description: Null when viewed through platform-admin oversight.
        memberCount:
          type: integer
        joinCode:
          type: string
          description: Visible to AGENT and above only.
    ClubList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Club'
    CreateClubRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        description:
          type: string
          maxLength: 500
    UpdateClubRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        description:
          type: string
          nullable: true
          maxLength: 500
    TransferOwnershipRequest:
      type: object
      additionalProperties: false
      required: [userId]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
    JoinClubRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          description: Club join code (8 chars) or invite code (12 chars); case-insensitive.
    Member:
      type: object
      additionalProperties: false
      required: [userId, username, role, status, joinedAt]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        role:
          $ref: '#/components/schemas/ClubRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
        joinedAt:
          $ref: '#/components/schemas/Timestamp'
    MemberPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        nextCursor:
          type: string
          nullable: true
    UpdateMemberRequest:
      type: object
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [ADMIN, AGENT, MEMBER]
        status:
          type: string
          enum: [ACTIVE, BANNED]
    Invite:
      type: object
      additionalProperties: false
      required: [id, clubId, role, maxUses, useCount, expiresAt, revokedAt, createdAt, status]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        clubId:
          $ref: '#/components/schemas/Uuid'
        role:
          type: string
          enum: [MEMBER, AGENT]
        maxUses:
          type: integer
        useCount:
          type: integer
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        status:
          type: string
          enum: [ACTIVE, EXPIRED, EXHAUSTED, REVOKED]
    InviteCreated:
      type: object
      additionalProperties: false
      required: [invite, code]
      properties:
        invite:
          $ref: '#/components/schemas/Invite'
        code:
          type: string
          description: Plaintext invite code, returned only once.
    InviteList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Invite'
    CreateInviteRequest:
      type: object
      additionalProperties: false
      properties:
        role:
          type: string
          enum: [MEMBER, AGENT]
          default: MEMBER
        maxUses:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1
        expiresInHours:
          type: integer
          minimum: 1
          maximum: 720
          default: 72
    AuditRecord:
      type: object
      additionalProperties: false
      required:
        [
          id,
          actorUserId,
          actorUsername,
          clubId,
          action,
          objectType,
          objectId,
          before,
          after,
          requestId,
          createdAt,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        actorUserId:
          type: string
          format: uuid
          nullable: true
        actorUsername:
          type: string
          nullable: true
        clubId:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        objectType:
          type: string
        objectId:
          type: string
          nullable: true
        before:
          type: object
          additionalProperties: true
          nullable: true
        after:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          nullable: true
        createdAt:
          $ref: '#/components/schemas/Timestamp'
    AuditPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditRecord'
        nextCursor:
          type: string
          nullable: true
...
    HandSummary:
      type: object
      description: No additionalProperties restriction because HandDetail extends it (allOf).
      required:
        [
          id,
          tableId,
          tableName,
          clubId,
          clubName,
          handNo,
          status,
          smallBlind,
          bigBlind,
          board,
          pot,
          playerCount,
          myNet,
          startedAt,
          endedAt,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        tableId:
          $ref: '#/components/schemas/Uuid'
        tableName:
          type: string
        clubId:
          $ref: '#/components/schemas/Uuid'
        clubName:
          type: string
        handNo:
          type: integer
          format: int64
        status:
          type: string
          enum: [COMPLETED, VOIDED]
        smallBlind:
          $ref: '#/components/schemas/ChipAmount'
        bigBlind:
          $ref: '#/components/schemas/ChipAmount'
        board:
          type: array
          items:
            $ref: './realtime.yaml#/components/schemas/Card'
        pot:
          $ref: '#/components/schemas/ChipAmount'
        playerCount:
          type: integer
        myNet:
          type: integer
          format: int64
          nullable: true
          description: The viewer's result; null when the viewer did not play or the hand was voided.
        startedAt:
          $ref: '#/components/schemas/Timestamp'
        endedAt:
          $ref: '#/components/schemas/Timestamp'
    HandSummaryPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HandSummary'
        nextCursor:
          type: string
          nullable: true
    HandParticipant:
      type: object
      additionalProperties: false
      required:
        [
          seat,
          userId,
          username,
          startingStack,
          endingStack,
          contributed,
          won,
          net,
          folded,
          shownCards,
        ]
      properties:
        seat:
          type: integer
        userId:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        startingStack:
          $ref: '#/components/schemas/ChipAmount'
        endingStack:
          type: integer
          format: int64
          nullable: true
        contributed:
          type: integer
          format: int64
          nullable: true
        won:
          type: integer
          format: int64
          nullable: true
        net:
          type: integer
          format: int64
          nullable: true
        folded:
          type: boolean
          nullable: true
        shownCards:
          type: array
          nullable: true
          description: Cards revealed at showdown (public); null when not shown.
          items:
            $ref: './realtime.yaml#/components/schemas/Card'
    HandEventRecord:
      type: object
      additionalProperties: false
      required: [seq, createdAt, event]
      properties:
        seq:
          type: integer
          format: int64
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        event:
          $ref: './realtime.yaml#/components/schemas/TableEventPayload'
    HandDetail:
      allOf:
        - $ref: '#/components/schemas/HandSummary'
        - type: object
          required:
            [buttonSeat, deckCommitment, voidReason, players, myHoleCards, events, viewerRole]
          properties:
            buttonSeat:
              type: integer
            deckCommitment:
              type: string
              description: SHA-256 commitment to the shuffled deck (internal audit).
            voidReason:
              type: string
              nullable: true
            players:
              type: array
              items:
                $ref: '#/components/schemas/HandParticipant'
            myHoleCards:
              type: array
              nullable: true
              description: The viewer's own hole cards; null for non-participants.
              items:
                $ref: './realtime.yaml#/components/schemas/Card'
            events:
              type: array
              description: Public action log of the hand (never contains unrevealed cards).
              items:
                $ref: '#/components/schemas/HandEventRecord'
            viewerRole:
              type: string
              enum: [PARTICIPANT, CLUB_STAFF, PLATFORM_ADMIN]
...
    AdminClub:
      type: object
      additionalProperties: false
      required: [id, name, status, ownerUserId, ownerUsername, memberCount, openTables, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        status:
          $ref: '#/components/schemas/ClubStatus'
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        ownerUsername:
          type: string
        memberCount:
          type: integer
        openTables:
          type: integer
        createdAt:
          $ref: '#/components/schemas/Timestamp'
...
    UpdateAccountStatusRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        reason:
          type: string
          minLength: 3
          maxLength: 500
          description: Recorded in the audit log.
...
    ErrorBody:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable summary (not for programmatic use).
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'